Clippy clean with -D warnings (default + output features)
* nsi-sys: allow `enum_variant_names` (bindgen output mirrors C header).
* nsi-ffi-wrap/build.rs: inline format arg in `eprintln!`.
* nsi-ffi-wrap/output/mod.rs: allow `redundant_allocation` and `new_ret_no_self`
  at the module root -- the triple-Box pattern and `*mut`-returning constructors
  are part of the ndspy C-API contract, not idiomatic Rust to clean up.
* nsi-jupyter: `std::u16::MAX` -> `u16::MAX` (legacy numeric constant).
* tests/test_utils.rs: replace manual elementwise loop with `copy_from_slice`.
* tests/test_box_layers.rs: allow type complexity for the deliberate triple-Box
  pointer-shape verification.
* tests/common/mod.rs: `#[allow(dead_code)]` on `add_metal_material` /
  `add_ground_plane` (test scaffolding kept for future use).
* tests/{materials,safety,geometry,test_utils,generate_test_image}: pick up
  `cargo clippy --fix` auto-suggestions (redundant imports, ref pattern,
  redundant closure, unnecessary cast, unused use).

+++ crates/nsi-ffi-wrap/tests/safety.rs
@@ -3,7 +3,6 @@
 //! These tests are designed to catch memory safety issues, particularly
 //! around callbacks, FFI boundaries, and lifetime management.
 
-use bytemuck;
 use nsi_ffi_wrap as nsi;
 use std::{
     panic::{self, AssertUnwindSafe},
@@ -175,7 +174,7 @@ fn thread_safety() {
                 ctx_clone.create(&node_name, nsi::ATTRIBUTES, None);
                 ctx_clone.set_attribute(
                     &node_name,
-                    &[nsi::i32!("thread_id", i as i32)],
+                    &[nsi::i32!("thread_id", i)],
                 );
             })
         })
